\name{senesceR-package}
\alias{senesceR-package}
\alias{senesceR}
\docType{package}
\title{
Simulation of an in vitro population of cells undergoing stochastic transition from replication to senescence.
}
\description{
Simulates inoculating a small population of cells into a flask of limited capacity.  Cells divide stochastically until flask becomes confluent, whereupon a sample of cells is transferred to a fresh flask and division can continue.  Cells can divide according to the commitment theory of cellular senescence (Kirkwood and Holliday) or just have fixed (or distributed) division potentials.
}
\details{
\tabular{ll}{
Package: \tab senesceR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-06-03\cr
License: \tab Artistic-2.0\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Conor Lawless

Maintainer: Conor Lawless <conor.lawless@ncl.ac.uk>
}
\references{

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
